+++ core/src/main/java/de/brainstormsoftworks/taloonerrl/system/InputProcessSystem.java
@@ -17,8 +17,8 @@ import com.artemis.systems.EntityProcessingSystem;
 import de.brainstormsoftworks.taloonerrl.components.FacingComponent;
 import de.brainstormsoftworks.taloonerrl.components.PlayerComponent;
 import de.brainstormsoftworks.taloonerrl.components.PositionComponent;
-import de.brainstormsoftworks.taloonerrl.core.Direction;
 import de.brainstormsoftworks.taloonerrl.core.engine.ComponentMappers;
+import de.brainstormsoftworks.taloonerrl.core.engine.Direction;
 import de.brainstormsoftworks.taloonerrl.core.engine.InputSystem;
 import de.brainstormsoftworks.taloonerrl.render.Renderer;
